Implementare il prompt Fixes #9
